test: Update settings tests to use the new app views
